fix(procmond): Disable network connections collection in compatibility tests
This commit updates the `test_platform_feature_compatibility` function in the OS compatibility tests to disable the collection of network connections. This change is made to prevent potential performance issues, as collecting network connections can be expensive. The adjustment aims to enhance the efficiency of the tests while maintaining the integrity of the compatibility checks.

Signed-off-by: UncleSp1d3r <unclespid3d@evilbitlabs.io>
Signed-off-by: UncleSp1d3r <unclesp1d3r@evilbitlabs.io>